* commit 'f8b0b53769e07caa9605d63b2a321589665b0a19': (177731 commits)
  pcie: merge s32v specific support to kernel 4.14
  s32r45x.dts: Change initrd offset to match uboot
  s32g275.dts: Change initrd offset to match uboot
  s32v344.dts: Change initrd offset to match uboot
  tmu: Update calibration table to resolve issues at 0 degrees or below
  linflex: Don't disable tx and rx during shutdown
  dts: Add new swt device tree node
  swt: Add support multi watchdog
  The fb driver was expecting pixclock from dts file to be in ps
  dts: sbc: Updated dts file to provide pixclock in Hz
  dts: pcie: Updated dts file to provide pixclock in Hz
  dts: evb28899: Updated dts file to provide pixclock in Hz
  dts: evb: Updated the dts file to provide pixclock in Hz
  dts: Correct clock source for swt
  fsl-edma: Resolve misleading identation warning
  clk: s32gen1: Remove clock gating
  drivers: fsl-dspi: Add "fsl,s32gen1-dspi"
  drivers: vf_pit: Add "fsl,s32gen1-pit" compatible
  drivers: fsl-stm: Add "fsl,s32gen1-stm" compatible
  dts: s32gen1: Update eDMA clock
  ...

+++ arch/metag/include/asm/checksum.h
@@ -1,3 +1,4 @@
+/* SPDX-License-Identifier: GPL-2.0 */
 #ifndef _METAG_CHECKSUM_H
 #define _METAG_CHECKSUM_H
 
@@ -59,8 +60,7 @@ extern __sum16 ip_fast_csum(const void *iph, unsigned int ihl);
  * returns a 16-bit checksum, already complemented
  */
 static inline __wsum csum_tcpudp_nofold(__be32 saddr, __be32 daddr,
-					unsigned short len,
-					unsigned short proto,
+					__u32 len, __u8 proto,
 					__wsum sum)
 {
 	unsigned long len_proto = (proto + len) << 8;
@@ -78,8 +78,8 @@ static inline __wsum csum_tcpudp_nofold(__be32 saddr, __be32 daddr,
 }
 
 static inline __sum16
-csum_tcpudp_magic(__be32 saddr, __be32 daddr, unsigned short len,
-		  unsigned short proto, __wsum sum)
+csum_tcpudp_magic(__be32 saddr, __be32 daddr, __u32 len,
+		  __u8 proto, __wsum sum)
 {
 	return csum_fold(csum_tcpudp_nofold(saddr, daddr, len, proto, sum));
 }
